convert example to TS
