+++ lib/src/interfaces/NotificationCategory.ts
@@ -13,14 +13,16 @@ export interface NotificationTextInput {
   placeholder: string;
 }
 
+export type ActivationMode = 'background' | 'foreground' | 'authenticationRequired' | 'destructive';
+
 export class NotificationAction {
   identifier: string;
-  activationMode: 'background' | 'foreground' | 'authenticationRequired' | 'destructive';
+  activationMode: ActivationMode;
   title: string;
   authenticationRequired: boolean;
   textInput?: NotificationTextInput;
 
-  constructor(identifier: string, activationMode: 'foreground' | 'authenticationRequired' | 'destructive', title: string, authenticationRequired: boolean, textInput?: NotificationTextInput) {
+  constructor(identifier: string, activationMode: ActivationMode, title: string, authenticationRequired: boolean, textInput?: NotificationTextInput) {
     this.identifier = identifier;
     this.activationMode = activationMode;
     this.title = title;
